api/requirements.txt
Flask
PyJWT
pika
